+++ src/main/java/com/ketheroth/uncrafter/common/inventory/container/AdvancedUncrafterContainer.java
@@ -1,16 +1,12 @@
 package com.ketheroth.uncrafter.common.inventory.container;
 
-import com.ketheroth.uncrafter.common.config.Configuration;
 import com.ketheroth.uncrafter.core.registry.UncrafterContainerTypes;
 import mcp.MethodsReturnNonnullByDefault;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.inventory.container.Container;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.IRecipe;
-import net.minecraft.item.crafting.IRecipeType;
 import net.minecraft.item.crafting.RecipeManager;
 import net.minecraft.util.Tuple;
 import net.minecraft.util.math.BlockPos;
@@ -19,44 +15,43 @@ import net.minecraftforge.items.SlotItemHandler;
 import net.minecraftforge.items.wrapper.InvWrapper;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class UncrafterContainer extends Container implements IUncrafterContainer {
+public class AdvancedUncrafterContainer extends Container implements IUncrafterContainer {
 
 	private final BlockPos pos;
 	private final PlayerEntity player;
 	private final IItemHandler playerInventory;
 	private final OutputHandler outputItems;
 	private final InputHandler inputItems;
+	private final EnchantmentHandler enchantmentHandler;
 	private Tuple<ItemStack, List<ItemStack>> cache = new Tuple<>(ItemStack.EMPTY, new ArrayList<>());
 
-	public UncrafterContainer(int windowId, PlayerInventory playerInventory, PlayerEntity player, BlockPos pos) {
-		super(UncrafterContainerTypes.UNCRAFTER_CONTAINER.get(), windowId);
+	public AdvancedUncrafterContainer(int windowId, PlayerInventory playerInventory, PlayerEntity player, BlockPos pos) {
+		super(UncrafterContainerTypes.ADVANCED_UNCRAFTER_CONTAINER.get(), windowId);
 		this.pos = pos;
 		this.player = player;
 		this.playerInventory = new InvWrapper(playerInventory);
 		this.outputItems = new OutputHandler(9, this);
 		this.inputItems = new InputHandler(1, this);
-
+		this.enchantmentHandler = new EnchantmentHandler(6);
 		//layout player inventory
 		for (int y = 0; y < 3; y++) {
 			for (int x = 0; x < 9; x++) {
-				this.addSlot(new SlotItemHandler(this.playerInventory, 9 + 9 * y + x, 8 + 18 * x, 84 + 18 * y));
+				addSlot(new SlotItemHandler(this.playerInventory, 9 + 9 * y + x, 8 + 18 * x, 84 + 18 * y));
 			}
 		}
 		for (int i = 0; i < 9; i++) {
-			this.addSlot(new SlotItemHandler(this.playerInventory, i, 8 + 18 * i, 142));
+			addSlot(new SlotItemHandler(this.playerInventory, i, 8 + 18 * i, 142));
 		}
 		//layout output inventory
 		for (int y = 0; y < 3; y++) {
 			for (int x = 0; x < 3; x++) {
-				this.addSlot(new SlotItemHandler(this.outputItems, 3 * y + x, 93 + 18 * x, 17 + 18 * y) {
+				addSlot(new SlotItemHandler(this.outputItems, 3 * y + x, 70 + 18 * x, 17 + 18 * y) {
 					@Override
 					public boolean mayPlace(@Nonnull ItemStack stack) {
 						return false;
@@ -65,12 +60,23 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 			}
 		}
 		//layout input inventory
-		this.addSlot(new SlotItemHandler(this.inputItems, 0, 35, 35) {
+		addSlot(new SlotItemHandler(this.inputItems, 0, 12, 35) {
 			@Override
 			public boolean mayPickup(PlayerEntity playerIn) {
-				return !UncrafterContainer.this.outputItems.isExtracting() && super.mayPickup(playerIn);
+				return !AdvancedUncrafterContainer.this.isInputLocked() && super.mayPickup(playerIn);
 			}
 		});
+
+		for (int y = 0; y < 3; y++) {
+			for (int x = 0; x < 2; x++) {
+				addSlot(new SlotItemHandler(this.enchantmentHandler, 2 * y + x, 134 + 18 * x, 17 + 18 * y) {
+					@Override
+					public boolean mayPlace(@Nonnull ItemStack stack) {
+						return false;
+					}
+				});
+			}
+		}
 	}
 
 	@Override
@@ -83,10 +89,12 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 		if (player instanceof ServerPlayerEntity) {
 			ItemStack itemstack = this.inputItems.extractItem(0, 64, false);
 			if (!itemstack.isEmpty()) {
-				if (player.isAlive() && !((ServerPlayerEntity) player).hasDisconnected()) {
-					player.inventory.placeItemBackInInventory(this.player.level, itemstack);
-				} else {
-					player.drop(itemstack, false);
+				if (!this.outputItems.isExtracting()) {
+					if (player.isAlive() && !((ServerPlayerEntity) player).hasDisconnected()) {
+						player.inventory.placeItemBackInInventory(this.player.level, itemstack);
+					} else {
+						player.drop(itemstack, false);
+					}
 				}
 			}
 		}
@@ -118,7 +126,12 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 
 	@Override
 	public boolean isInputLocked() {
-		return this.outputItems.isExtracting();
+		return this.outputItems.isExtracting() || this.enchantmentHandler.isExtracting();
+	}
+
+	@Override
+	public OutputHandler getOutputHandler() {
+		return this.outputItems;
 	}
 
 	@Override
@@ -127,14 +140,8 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 	}
 
 	@Override
-	public OutputHandler getOutputHandler() {
-		return this.outputItems;
-	}
-
-	@Nullable
-	@Override
 	public EnchantmentHandler getEnchantmentHandler() {
-		return null;
+		return this.enchantmentHandler;
 	}
 
 	@Override
@@ -143,13 +150,13 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 	}
 
 	@Override
-	public void setCache(List<ItemStack> b) {
-		this.cache = new Tuple<>(this.cache.getA(), b);
+	public void setCache(ItemStack a, List<ItemStack> b) {
+		this.cache = new Tuple<>(a, b);
 	}
 
 	@Override
-	public void setCache(ItemStack a, List<ItemStack> b) {
-		this.cache = new Tuple<>(a, b);
+	public void setCache(List<ItemStack> b) {
+		this.cache = new Tuple<>(cache.getA(), b);
 	}
 
 	@Override
@@ -159,21 +166,7 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 
 	@Override
 	public boolean isAdvanced() {
-		return false;
-	}
-
-	@Nullable
-	public static IRecipe<?> searchRecipe(ItemStack input, RecipeManager recipeManager) {
-		Item inputItem = input.getItem();
-		Optional<IRecipe<?>> optionalRecipe = recipeManager.getRecipes().stream()
-				.filter(recipe -> recipe.getType().equals(IRecipeType.CRAFTING))
-				.filter(recipe -> !Configuration.BLACKLIST.get().contains(recipe.getId().toString()))
-				.filter(recipe -> !Configuration.IMC_BLACKLIST.contains(recipe.getId().toString()))
-				.filter(recipe -> recipe.canCraftInDimensions(3, 3)
-						&& recipe.getResultItem().getItem() == inputItem
-						&& !recipe.getIngredients().isEmpty())
-				.findAny();
-		return optionalRecipe.orElse(null);
+		return true;
 	}
 
 }