+++ src/main/java/com/ketheroth/uncrafter/common/inventory/container/EnchantmentHandler.java
@@ -1,0 +1,50 @@
+package com.ketheroth.uncrafter.common.inventory.container;
+
+import com.ketheroth.uncrafter.common.config.Configuration;
+import net.minecraft.item.ItemStack;
+import net.minecraftforge.items.ItemStackHandler;
+
+import javax.annotation.Nonnull;
+
+public class EnchantmentHandler extends ItemStackHandler {
+
+	private int extracted;
+
+	public EnchantmentHandler(int size) {
+		super(size);
+		this.extracted = 0;
+	}
+
+	@Nonnull
+	@Override
+	public ItemStack extractItem(int slot, int amount, boolean simulate) {
+		ItemStack extractedStack = super.extractItem(slot, amount, simulate);
+		if (!simulate) {
+			extracted++;
+		}
+		if (extracted >= Configuration.ENCHANTMENT_EXTRACT_AMOUNT.get() || this.isEmpty()) {
+			for (int i = 0; i < this.getSlots(); i++) {
+				this.setStackInSlot(i, ItemStack.EMPTY);
+			}
+		}
+		return extractedStack;
+	}
+
+	private boolean isEmpty() {
+		for (int i = 0; i < this.getSlots(); i++) {
+			if (!this.getStackInSlot(i).isEmpty()) {
+				return false;
+			}
+		}
+		return true;
+	}
+
+	public void resetExtracting() {
+		this.extracted = 0;
+	}
+
+	public boolean isExtracting() {
+		return this.extracted > 0;
+	}
+
+}