+++ src/main/java/com/ketheroth/uncrafter/core/registry/UncrafterBlocks.java
@@ -1,6 +1,7 @@
 package com.ketheroth.uncrafter.core.registry;
 
 import com.ketheroth.uncrafter.Uncrafter;
+import com.ketheroth.uncrafter.common.bock.AdvancedUncrafterBlock;
 import com.ketheroth.uncrafter.common.bock.UncrafterBlock;
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
@@ -20,5 +21,8 @@ public class UncrafterBlocks {
 
 	public static final RegistryObject<Block> UNCRAFTER_BLOCK = BLOCKS.register("uncrafter", () -> new UncrafterBlock(AbstractBlock.Properties.of(Material.STONE, MaterialColor.SAND).requiresCorrectToolForDrops().strength(3.0F, 9.0F)));
 	public static final RegistryObject<Item> UNCRAFTER_ITEM = ITEMS.register("uncrafter", () -> new BlockItem(UNCRAFTER_BLOCK.get(), new Item.Properties().tab(ItemGroup.TAB_DECORATIONS)));
+
+	public static final RegistryObject<Block> ADVANCED_UNCRAFTER_BLOCK = BLOCKS.register("advanced_uncrafter", () -> new AdvancedUncrafterBlock(AbstractBlock.Properties.of(Material.STONE, MaterialColor.SAND).requiresCorrectToolForDrops().strength(3.0F, 9.0F)));
+	public static final RegistryObject<Item> ADVANCED_UNCRAFTER_ITEM = ITEMS.register("advanced_uncrafter", () -> new BlockItem(ADVANCED_UNCRAFTER_BLOCK.get(), new Item.Properties().tab(ItemGroup.TAB_DECORATIONS)));
 
 }