+++ src/main/java/com/ketheroth/uncrafter/core/registry/UncrafterContainerTypes.java
@@ -1,6 +1,7 @@
 package com.ketheroth.uncrafter.core.registry;
 
 import com.ketheroth.uncrafter.Uncrafter;
+import com.ketheroth.uncrafter.common.inventory.container.AdvancedUncrafterContainer;
 import com.ketheroth.uncrafter.common.inventory.container.UncrafterContainer;
 import net.minecraft.inventory.container.ContainerType;
 import net.minecraftforge.common.extensions.IForgeContainerType;
@@ -14,5 +15,8 @@ public class UncrafterContainerTypes {
 
 	public static final RegistryObject<ContainerType<UncrafterContainer>> UNCRAFTER_CONTAINER = CONTAINERS.register("uncrafter_container",
 			() -> IForgeContainerType.create((windowId, inv, data) -> new UncrafterContainer(windowId, inv, inv.player, data.readBlockPos())));
+
+	public static final RegistryObject<ContainerType<AdvancedUncrafterContainer>> ADVANCED_UNCRAFTER_CONTAINER = CONTAINERS.register("advanced_uncrafter_container",
+			() -> IForgeContainerType.create((windowId, inv, data) -> new AdvancedUncrafterContainer(windowId, inv, inv.player, data.readBlockPos())));
 
 }