+++ CHANGELOG.md
@@ -3,11 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 1.16.5-0.3.0
+
+### Added
+- Advanced Uncrafter. Allow three ingredients to be retrieved. Allow three enchantments to be retrieved (as enchanted books), but you can't retrieve a curse.
+- When extracting an ingredient from the output slots, the input slot is locked until you have taken all possible ingredients.
+  A little cross appears on the input slot when it is locked.
 
 ### Changed
-- Changed the texture again (it's better now).
+- Changed the texture for the uncrafter again (it's better now).
 
+### Fixed
+- Uncrafter having end stone as bottom texture instead of crying obsidian
 
 ## 1.16.5-0.2.0